+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Translated the CHANGELOG from Korean to English.
+- Converted enum values in `CashReceiptTradeOption` to snake case for consistency.
 
 ## [1.0.0] - 2025-04-11